pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pygentix"
version = "0.1.1"
description = "Composable AI agent framework — drop-in LLM tool-calling, structured output, and SQLAlchemy integration for any Python project."
authors = [
    { name = "Andre Perussi", email = "dlperussi@hotmail.com" },
]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Typing :: Typed",
]
dependencies = [
    "sqlalchemy>=2.0",
]

[project.optional-dependencies]
ollama = ["ollama>=0.4"]
openai = ["openai>=1.0"]
gemini = ["google-genai>=1.0"]
scheduler = ["croniter>=1.0"]
all = [
    "ollama>=0.4",
    "openai>=1.0",
    "google-genai>=1.0",
    "croniter>=1.0",
]
dev = [
    "ollama>=0.4",
    "openai>=1.0",
    "google-genai>=1.0",
    "croniter>=1.0",
    "pymupdf>=1.24",
    "pytest>=8.0",
]

[project.urls]
Homepage = "https://github.com/andreperussi/pygentix"
Repository = "https://github.com/andreperussi/pygentix"
Issues = "https://github.com/andreperussi/pygentix/issues"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
pygentix = ["py.typed"]

[tool.pytest.ini_options]
testpaths = ["tests"]
